--- server/odc-common/src/main/java/com/oceanbase/odc/common/config/EncryptableConfigurations.java
@@ -1,190 +1,0 @@
-/*
- * Copyright (c) 2023 OceanBase.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.oceanbase.odc.common.config;
-
-import static java.util.regex.Pattern.CASE_INSENSITIVE;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Properties;
-import java.util.Set;
-import java.util.regex.Pattern;
-
-import org.apache.commons.configuration2.PropertiesConfiguration;
-import org.apache.commons.configuration2.PropertiesConfigurationLayout;
-import org.apache.commons.configuration2.ex.ConfigurationException;
-
-import com.oceanbase.odc.common.crypto.Encryptors;
-import com.oceanbase.odc.common.crypto.TextEncryptor;
-import com.oceanbase.odc.common.util.StringUtils;
-
-import lombok.extern.slf4j.Slf4j;
-
-/**
- * 支持配置项值加密的配置读取，密钥维护在环境变量，<br>
- * ACI_VAR_ 前缀的环境变量用于适配 antcode ACI 流水线环境。
- */
-@Slf4j
-public class EncryptableConfigurations {
-    private static final Pattern ENCRYPT_KEY_PATTERN =
-            Pattern.compile(".*(password|username|host|port|commandline|key|secret|jdbcUrl|endpoint|authorization).*",
-                    CASE_INSENSITIVE);
-
-    private static final String ENCRYPTED_PREFIX = "ENC@";
-    private static final String DOT_ENV_FILE = "../../.env";
-
-    private static final EncryptablePropertyDetector encryptableDetector;
-    private static final TextEncryptor valueEncryptor;
-
-    static {
-        encryptableDetector = new EncryptablePropertyDetector();
-        valueEncryptor = Encryptors.blowFishZeroPaddingBase64(new SecretGetter().secret());
-    }
-
-    public static Map<String, String> loadProperties(String path) {
-        encryptFileIfRequires(path);
-
-        File file = new File(path);
-        PropertiesConfiguration config = new PropertiesConfiguration();
-        PropertiesConfigurationLayout layout = new PropertiesConfigurationLayout();
-        try {
-            layout.load(config, new FileReader(file));
-        } catch (ConfigurationException | FileNotFoundException e) {
-            throw new RuntimeException("load properties file failed:", e);
-        }
-        Map<String, String> properties = new HashMap<>();
-        Set<String> keys = layout.getKeys();
-        for (String key : keys) {
-            String value = config.getProperty(key).toString();
-            value = decryptIfRequired(value);
-            properties.put(key, value);
-        }
-        return properties;
-    }
-
-    private static String decryptIfRequired(String value) {
-        if (!encryptableDetector.isEncrypted(value)) {
-            return value;
-        }
-        return valueEncryptor.decrypt(encryptableDetector.unwrapEncryptedValue(value));
-    }
-
-    public static void encryptFileIfRequires(String fileName) {
-        File file = new File(fileName);
-        PropertiesConfiguration config = new PropertiesConfiguration();
-        PropertiesConfigurationLayout layout = new PropertiesConfigurationLayout();
-        try {
-            layout.load(config, new FileReader(file));
-        } catch (ConfigurationException | FileNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-
-        Set<String> keys = layout.getKeys();
-        boolean detectUnencryptedValue = false;
-        for (String key : keys) {
-            String value = config.getProperty(key).toString();
-            if (ENCRYPT_KEY_PATTERN.matcher(key).matches()
-                    && StringUtils.isNotBlank(value)
-                    && !encryptableDetector.isEncrypted(value)) {
-                String encryptedValue = ENCRYPTED_PREFIX + valueEncryptor.encrypt(value);
-                config.setProperty(key, encryptedValue);
-                detectUnencryptedValue = true;
-            }
-        }
-        if (detectUnencryptedValue) {
-            log.info("detect unencrypted value, encrypt and save it, fileName={}", fileName);
-            try {
-                layout.save(config, new FileWriter(file));
-            } catch (ConfigurationException | IOException e) {
-                throw new RuntimeException("save properties file failed:", e);
-            }
-        }
-    }
-
-    private static class EncryptablePropertyDetector {
-
-        public boolean isEncrypted(String value) {
-            if (value != null) {
-                return value.startsWith(ENCRYPTED_PREFIX);
-            }
-            return false;
-        }
-
-        public String unwrapEncryptedValue(String value) {
-            return value.substring(ENCRYPTED_PREFIX.length());
-        }
-    }
-
-    private static class SecretGetter {
-        private static final String SECRET_ENV_KEY = "ODC_CONFIG_SECRET";
-        private static final String SECRET_ENV_ACI_KEY = "ACI_VAR_ODC_CONFIG_SECRET";
-        private final Properties dotEnvProperties;
-
-        SecretGetter() {
-            dotEnvProperties = dotEnvProperties();
-        }
-
-        public String secret() {
-            String odcConfigSecret = getSystemProperty(SECRET_ENV_KEY);
-            if (StringUtils.isNotBlank(odcConfigSecret)) {
-                return odcConfigSecret;
-            }
-            odcConfigSecret = getSystemProperty(SECRET_ENV_ACI_KEY);
-            if (StringUtils.isNotBlank(odcConfigSecret)) {
-                return odcConfigSecret;
-            }
-            throw new RuntimeException("environment variable 'ODC_CONFIG_SECRET' not set");
-        }
-
-        private Properties dotEnvProperties() {
-            Properties properties = new Properties();
-            File file = new File(DOT_ENV_FILE);
-            if (file.exists()) {
-                try (FileInputStream is = new FileInputStream(file)) {
-                    properties.load(is);
-                } catch (IOException e) {
-                    log.warn("load .env failed, reason={}", e.getMessage());
-                }
-            } else {
-                log.info("skip load due .env file not exists");
-            }
-            return properties;
-        }
-
-        private String getSystemProperty(String key) {
-            String value = System.getProperty(key);
-            if (StringUtils.isNotBlank(value)) {
-                return value;
-            }
-            value = System.getenv(key);
-            if (StringUtils.isNotBlank(value)) {
-                return value;
-            }
-            value = dotEnvProperties.getProperty(key);
-            if (StringUtils.isNotBlank(value)) {
-                return value;
-            }
-            return null;
-        }
-    }
-
-}
