--- server/odc-common/src/test/resources/test-encrypt-configuration.properties
@@ -1,2 +1,0 @@
-key1=ENC@TZiDnrN0hNw=
-key2=ENC@p70OxV+YLw0=
