+++ server/odc-test/src/test/java/com/oceanbase/odc/test/tool/EncryptableConfigurationsTest.java
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.common.config;
+
+package com.oceanbase.odc.test.tool;
+
+import static org.junit.Assert.*;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -21,9 +24,10 @@ import java.util.Map;
 import org.junit.Assert;
 import org.junit.Test;
 
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
+/**
+ * @author gaoda.xy
+ * @date 2023/9/28 10:24
+ */
 public class EncryptableConfigurationsTest {
     private static final String TEST_CONFIG_FILE = "src/test/resources/test-encrypt-configuration.properties";
 
@@ -32,9 +36,10 @@ public class EncryptableConfigurationsTest {
         Map<String, String> load = EncryptableConfigurations.loadProperties(TEST_CONFIG_FILE);
 
         Map<String, String> expected = new HashMap<>();
-        expected.put("key1", "value1");
-        expected.put("key2", "value2");
+        expected.put("key1", "oceanbase developer center");
+        expected.put("key2", "build the best database develop platform");
 
         Assert.assertEquals(expected, load);
     }
+
 }
