+++ server/odc-test/src/test/resources/test-encrypt-configuration.properties
@@ -1,0 +1,18 @@
+#
+# Copyright (c) 2023 OceanBase.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+key1=ENC@EpRFrpc+NKGHYedqceH37TNfCCtmd1Sqz9koiTRN0XYnp0TfzZln2m2lGpsD6e6H
+key2=ENC@v9RfPncKuNNqOISn7KZUvmFqCJGfeDO/frHnX61gTcaqozhqCYNVkTDkoRXmFbUjMjHvV2EKr1hpaGDgDsWrjQ==
